refactor: rename check files
