+++ src/helpers/dataCheck/registrationCheck.ts
